First cut.
